// src/preview/steingardur/photos.gen.ts
// GENERATED by _docs/steingardur-harvest-2026-10-09/build_photos.py: served sizes (width, height) of public/steingardur/<id>-1920.jpg
export const SIZE: Record<string, [number, number]> = {
  'hitachi-gata': [960, 720],
  'hitachi-gardur': [960, 720],
  'mini-kef': [960, 720],
  'kobelco-kefli': [960, 720],
  'heklureitur': [1920, 1440],
  'bogdan-runar': [1920, 1440],
  'heklureitur-tveir': [1536, 2048],
  'heklureitur-lagnir': [1536, 2048],
  'kopavogur': [1920, 1440],
  'fleygun-komatsu': [1920, 1440],
  'fleygun-hjolagrafa': [1536, 2048],
  'nautholsvegur': [1536, 2048],
  'flutningur': [720, 540],
  'jcb-merkt': [1920, 1440],
  'snjo-gata': [1920, 1440],
  'snjo-nott': [1440, 1080],
  'snjo-tvaer': [1920, 1920],
  'snjo-jcb': [1920, 1440],
  'snjo-klefi': [1920, 1422],
  'ljos-akur': [960, 720],
  'ljos-skurdur': [720, 960],
  'ljos-kefli': [960, 720],
  'ljos-skapur': [960, 720],
  'ljos-gata': [960, 720],
  'ljos-splaes': [1440, 1080],
  'hitachi-turn': [960, 720],
  'hitachi-vegur': [960, 720],
  'jcb-hjol': [960, 720],
  'kobelco-gras': [960, 720],
  'scania': [1280, 960],
  'scania-vagn': [960, 720],
  'hopur': [1920, 1440],
  'hitaveita': [1920, 1440],
  'beltagrafa': [1920, 1440],
  'skoli': [1920, 1081],
  'dofrahella': [1214, 623],
  'lagnir-rokkur': [1920, 1080],
  'mini-dusk': [1440, 1080],
  'kobelco-malbik': [1920, 1440],
  'gatnagerd': [1920, 1440],
  'stett': [1920, 1440],
}